// src/app/core/models/policyDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface PolicyDto {
  id: string;
  tenantId: string;
  name: string;
  policyNumber: string;
  description: string;
  payoutAmount: number;
  coverageAmount: number;
  price: number;
  waitingPeriodDays: number;
  maxClaimAmount: number;
  isActive: boolean;
  status: string;
  startDate: string;
  endDate: string;
  premiumAmount: number;
  createdAt: string;
  updatedAt: string;
  createdBy: string;
  updatedBy: string;
}