// src/app/core/models/tenantCreateUpdateDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface TenantCreateUpdateDto {
  id: string;
  email: string;
  password: string;
  name: string;
  domain: string;
  address: string;
  phone1: string;
  phone2: string;
  registrationNumber: string;
  type: string;
  subscriptionPlanId: string;
}